requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.0
spacy==3.7.4
nltk==3.8.1
google-generativeai==0.7.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
pydantic==2.7.3
httpx==0.27.0
aiofiles==23.2.1
PyPDF2==3.0.1
pdf2image==1.17.0
pyspellchecker==0.8.1
